--- stagecoach/price-pusher/price_pusher/core/listeners/__init__.py
@@ -1,3 +1,0 @@
-from .listener import PriceListener
-
-__all__ = ["PriceListener"]
