+++ stagecoach/price-pusher/price_pusher/core/request_handlers/chain.py
@@ -2,15 +2,15 @@ from typing import Optional
 
 from pragma.core.assets import PragmaAsset
 from pragma.core.entry import Entry
-from pragma.publisher.client import PragmaAPIClient
+from pragma.publisher.client import PragmaOnChainClient
 
-from price_pusher.core.listeners.listener import IRequestHandler
+from price_pusher.core.request_handlers.interface import IRequestHandler
 
 
-class APIRequestHandler(IRequestHandler):
-    client: PragmaAPIClient
+class ChainRequestHandler(IRequestHandler):
+    client: PragmaOnChainClient
 
-    def __init__(self, client: PragmaAPIClient) -> None:
+    def __init__(self, client: PragmaOnChainClient) -> None:
         self.client = client
 
     async def fetch_latest_asset_price(self, asset: PragmaAsset) -> Optional[Entry]:
