+++ stagecoach/price-pusher/price_pusher/main.py
@@ -7,14 +7,17 @@ from typing import Optional, List
 from pragma.publisher.client import FetcherClient
 
 from price_pusher.core.poller import PricePoller
-from price_pusher.core.listeners import PriceListener
-from price_pusher.core.listeners.request_handlers import (
+from price_pusher.core.listener import PriceListener
+from price_pusher.core.request_handlers import (
     APIRequestHandler,
     ChainRequestHandler,
 )
 from price_pusher.core.pusher import PricePusher
-from price_pusher.fetchers import add_all_fetchers
-from price_pusher.configs.price_config import PriceConfig
+from price_pusher.core.fetchers import add_all_fetchers
+from price_pusher.configs.price_config import (
+    PriceConfig,
+    get_all_unique_assets_from_config_list,
+)
 from price_pusher.configs.cli import setup_logging, load_private_key, create_client
 from price_pusher.orchestrator import Orchestrator
 
@@ -60,7 +63,7 @@ async def main(
     listener = PriceListener(
         request_handler=RequestHandlerClass(client=pragma_client.client),
         polling_frequency_in_s=2,
-        assets=[],
+        assets=get_all_unique_assets_from_config_list(price_configs),
     )
     pusher = PricePusher(client=pragma_client)
     orchestrator = Orchestrator(poller=poller, listener=listener, pusher=pusher)
